Seperate declaration and implementation -- for long term maintenance
